+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added code-snippet for producing a boilerplate Storybook component story.
 - Added new icon component `<Close>`.
 - Added new component `<ButtonClose>`.
+- Added new component `<Modal>`.
 
 ### Changes